#15 refactor: RadioButtonGroup 컴포넌트를 components 파일로 이동
