+++ src/pages/PartySurveyForm/components/PartyConditionForm.module.scss
@@ -48,31 +48,6 @@
     line-height: rem(1.7);
   }
 
-  .RadioGroupRoot {
-    display: flex;
-    gap: rem(8px);
-    width: 100%;
-  }
-
-  .RadioGroupItem {
-    border: 1px solid #e5e5e5;
-    flex: 1;
-    height: rem(44px);
-    border-radius: rem(8px);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    color: #777777;
-    font-size: rem(14px);
-    line-height: rem(1.7);
-    margin-bottom: rem(60px);
-  }
-
-  .RadioGroupItem[data-state='checked'] {
-    color: #333333;
-    border-color: #333333;
-  }
-
   .active {
     color: #333333;
     border-color: #333333;
